+++ src/lua/CommunityBalanceMod/AdvancedPrototypelab/NS2Utility.lua
@@ -1,6 +1,6 @@
 
 
-oldBuildClassToGrid = BuildClassToGrid
+local oldBuildClassToGrid = BuildClassToGrid
 function BuildClassToGrid()
     local ClassToGrid = oldBuildClassToGrid()
     ClassToGrid["AdvancedPrototypeLab"] = { 4, 5 }