+++ src/lua/CommunityBalanceMod/FortressPvE/Minimap/Globals.lua
@@ -1,0 +1,7 @@
+Script.Load("lua/CommunityBalanceMod/Scripts/EnumUtils.lua")
+
+EnumUtils.AppendToEnum(kMinimapBlipType, "FortressCrag")
+EnumUtils.AppendToEnum(kMinimapBlipType, "FortressShift")
+EnumUtils.AppendToEnum(kMinimapBlipType, "FortressShade")
+EnumUtils.AppendToEnum(kMinimapBlipType, "FortressWhip")
+EnumUtils.AppendToEnum(kMinimapBlipType, "FortressWhipMature")