+++ src/lua/CommunityBalanceMod/FortressPvE/Shade.lua
@@ -149,6 +149,8 @@ if Server then
            -- self:SetTechId(kTechId.FortressShade)
             self:UpgradeToTechId(kTechId.FortressShade)
 
+            self:MarkBlipDirty()
+
             local techTree = self:GetTeam():GetTechTree()
             local researchNode = techTree:GetTechNode(kTechId.Shade)