+++ src/lua/CommunityBalanceMod/FortressPvE/Whip_Server.lua
@@ -155,3 +155,11 @@ function Whip:BileShower()
 
     return self.enervating and isAlive
 end
+
+
+function Whip:OnMaturityComplete()
+    if HasMixin(self, "MapBlip") then 
+        self:MarkBlipDirty()
+    end
+    self:GiveUpgrade(kTechId.WhipBombard)
+end