+++ src/lua/CommunityBalanceMod/Minimap/ARC_Server.lua
@@ -1,0 +1,12 @@
+
+function ARC:OnEntityChange(oldId)
+
+    if HasMixin(self, "MapBlip") then 
+        self:MarkBlipDirty()
+   end
+
+    if self.targetedEntity == oldId then
+        self.targetedEntity = Entity.invalidId
+    end    
+
+end